+++ doc/examples/add-job.py
@@ -77,6 +77,7 @@ except KeyError:
 # ------------------------------------------------------------
 
 inputcollection = client.new("dataCollection")
+initobj(inputcollection, jobdata['input'])
 
 for ds in jobdata['input']['datasets']:
     query = Query(client, "Dataset", conditions={
@@ -110,6 +111,7 @@ inputcollection.create()
 # ------------------------------------------------------------
 
 outputcollection = client.new("dataCollection")
+initobj(outputcollection, jobdata['output'])
 
 for ds in jobdata['output']['datasets']:
     query = Query(client, "Investigation", conditions={