+++ tests/test_05_setup.py
@@ -7,27 +7,45 @@ these investigations.  It then compares the result with a reference
 dump file.
 """
 
-import os.path
 import filecmp
+import os.path
+import re
 import yaml
 import pytest
 import icat
 import icat.config
 from icat.query import Query
-from conftest import getConfig, require_icat_version
-from conftest import gettestdata, callscript, filter_file, yaml_filter
+from conftest import (getConfig, icat_version,
+                      gettestdata, get_reference_dumpfile, callscript,
+                      filter_file, yaml_filter)
 
+
+# Study is broken in icat.server older then 4.6.0, see
+# icatproject/icat.server#155.  We do not maintain specific set of
+# test data for 4.6.  Therefore we skip testing Study for ICAT older
+# then 4.7.
+skip_study = icat_version < "4.7.0"
 
 testinput = gettestdata("example_data.yaml")
-refdump = gettestdata("icatdump.yaml")
+refdump = get_reference_dumpfile("yaml")
 users = [ "acord", "ahau", "jbotu", "jdoe", "nbour", "rbeck" ]
 refsummary = { "root": gettestdata("summary") }
 for u in users:
     refsummary[u] = gettestdata("summary.%s" % u)
 # Labels used in test dependencies.
-alldata = ["init", "sample_durol", "sample_nimnga", "sample_nio", "inv_081",
-           "inv_101", "inv_121", "invdata_081", "invdata_101", "invdata_121",
-           "job1", "rdf1", "study1", "pub1"]
+if not skip_study:
+    alldata = ["init", "sample_durol", "sample_nimnga", "sample_nio",
+               "inv_081", "inv_101", "inv_121",
+               "invdata_081", "invdata_101", "invdata_121",
+               "job1", "rdf1", "study1", "pub1"]
+else:
+    alldata = ["init", "sample_durol", "sample_nimnga", "sample_nio",
+               "inv_081", "inv_101", "inv_121",
+               "invdata_081", "invdata_101", "invdata_121",
+               "job1", "rdf1", "pub1"]
+summary_study_filter = (re.compile(r"^((?:Study(?:Investigation)?)\s*) : \d+$"),
+                        r"\1 : 0")
+
 
 @pytest.fixture(scope="module")
 def data():
@@ -162,6 +180,7 @@ def test_add_relateddatafile(data, user, rdfname):
                  marks=pytest.mark.dependency(
                      name="study1", depends=["inv_101", "inv_121"])),
 ])
+@pytest.mark.skipif(skip_study, reason="Issue icatproject/icat.server#155")
 def test_add_study(data, user, studyname):
     client, conf = getConfig(confSection=user)
     client.login(conf.auth, conf.credentials)
@@ -192,12 +211,10 @@ def test_add_publication(data, user, pubname):
     publication.investigation = client.assertedSearch(query)[0]
     publication.create()
 
-
 @pytest.mark.dependency(depends=alldata)
 def test_check_content(standardCmdArgs, tmpdirsec):
     """Dump the resulting content and compare with a reference dump.
     """
-    require_icat_version("4.6.0", "Issue icatproject/icat.server#155")
     dump = os.path.join(tmpdirsec, "dump.yaml")
     fdump = os.path.join(tmpdirsec, "dump-filter.yaml")
     reffdump = os.path.join(tmpdirsec, "dump-filter-ref.yaml")
@@ -213,6 +230,10 @@ def test_check_summary_root(standardCmdArgs, tmpdirsec):
     """
     summary = os.path.join(tmpdirsec, "summary")
     ref = refsummary["root"]
+    if skip_study:
+        reff = os.path.join(tmpdirsec, "summary-filter-ref")
+        filter_file(ref, reff, *summary_study_filter)
+        ref = reff
     with open(summary, "wt") as out:
         callscript("icatsummary.py", standardCmdArgs, stdout=out)
     assert filecmp.cmp(ref, summary), "ICAT content was not as expected"
@@ -227,6 +248,10 @@ def test_check_summary_user(tmpdirsec, user):
     """
     summary = os.path.join(tmpdirsec, "summary.%s" % user)
     ref = refsummary[user]
+    if skip_study:
+        reff = os.path.join(tmpdirsec, "summary-filter-ref.%s" % user)
+        filter_file(ref, reff, *summary_study_filter)
+        ref = reff
     _, conf = getConfig(confSection=user)
     with open(summary, "wt") as out:
         callscript("icatsummary.py", conf.cmdargs, stdout=out)